Rename views.tsx to overview.tsx
