+++ preview-src/app.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { useContext, useState, useEffect } from 'react';
 import { render } from 'react-dom';
-import { Overview } from './views';
-import PRContext from './context';
+import { Overview } from './overview';
+import PullRequestContext from './context';
 import { PullRequest } from './cache';
 
 export function main() {
@@ -12,7 +12,7 @@ export function main() {
 }
 
 function Root({ children }) {
-	const ctx = useContext(PRContext);
+	const ctx = useContext(PullRequestContext);
 	const [pr, setPR] = useState<PullRequest>(ctx.pr);
 	useEffect(() => {
 		ctx.onchange = setPR;
